backend/requirements.txt
# backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
pyjwt>=2.6.0
pycryptodome>=3.17.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
websockets>=11.0.0
cryptography>=41.0.0
pydantic>=1.10.0
bcrypt>=4.0.0
python-jose>=3.3.0
redis